Adds scenario for if a user does not want to install asdf on Ubuntu (#4)
Addresses #3

Hole in the test coverage! We are not testing if a user wishes to skip
the asdf install on this playbook (which would just do nothing to the
target instance). The ability to opt-out is a huge selling point for
eng. This adds an automated scenario for Linux. It does not cover macOS
because the macOS tests take almost an order of magnitude longer to run
in GitHub Actions, so leaving macOS off and considering it "good
enough", but the code path is different so there is some risk going
forward.